Feature: Implement POC for handling Object Variables inside Unreals property system
This commit introduces handling of object variables within unreal engines property system. It aims to implement an easy way of extending property types and registering it right with the property system of gorgeous things.

Files that are part of this POC are:
- UK2Node_MyCustomNode.cpp
- UK2Node_MyCustomNode.cpp
- GorgeousObjectVariableStructures.h

Files that were changed to comply with this feature change:
- GorgeousObjectVariableDefinitions.h
- GorgeousObjectVariableDefinitions.cpp